Refactoring Option into two files
